reading expiresIN as long

+++ app/src/main/java/com/diegorezm/ratemymusic/modules/spotify_auth/data/remote/models/SpotifyTokenDTO.kt
@@ -6,7 +6,7 @@ import com.google.gson.annotations.SerializedName
 data class SpotifyTokenDTO(
     @SerializedName("access_token") val accessToken: String,
     @SerializedName("token_type") val tokenType: String,
-    @SerializedName("expires_in") val expiresIn: Int,
+    @SerializedName("expires_in") val expiresIn: Long,
     @SerializedName("refresh_token") val refreshToken: String,
     @SerializedName("scope") val scope: String
 )
@@ -17,7 +17,7 @@ fun SpotifyTokenDTO.toEntity(): SpotifyTokenEntity {
     return SpotifyTokenEntity(
         accessToken = this.accessToken,
         tokenType = this.tokenType,
-        expiresIn = expiresIn.toLong(),
+        expiresIn = expiresIn,
         refreshToken = this.refreshToken,
         scope = this.scope
     )
